[PM-24721] Refactor AccountKeys to top-level common model
Refactored `AccountKeys` to be a top-level model `AccountKeysJson` in the network module.
This change allows `AccountKeysJson` to be used in `GetTokenResponseJson` and `SyncResponseJson.Profile`.

The corresponding test fixtures for creating mock `AccountKeysJson` instances were also moved to a new file `AccountKeysJsonUtil.kt` in the network testFixtures.

`AuthDiskSource` and its implementation `AuthDiskSourceImpl`, along with the test fake `FakeAuthDiskSource`, were updated to use the new `AccountKeysJson` model.
