+++ network/src/main/kotlin/com/bitwarden/network/model/SyncResponseJson.kt
@@ -177,12 +177,15 @@ data class SyncResponseJson(
         @SerialName("twoFactorEnabled")
         val isTwoFactorEnabled: Boolean,
 
-        @Deprecated("Use accountKeys instead", ReplaceWith("accountKeys"))
+        @Deprecated(
+            message = "Use `accountKeys` instead",
+            ReplaceWith("profile.accountKeys?.publicKeyEncryptionKeyPair?.wrappedPrivateKey"),
+        )
         @SerialName("privateKey")
         val privateKey: String?,
 
         @SerialName("accountKeys")
-        val accountKeys: AccountKeys?,
+        val accountKeys: AccountKeysJson?,
 
         @SerialName("premium")
         val isPremium: Boolean,
@@ -416,77 +419,6 @@ data class SyncResponseJson(
             @SerialName("managePolicies")
             val shouldManagePolicies: Boolean,
         )
-
-        /**
-         * Represents private keys in the vault response.
-         *
-         * @property signatureKeyPair The signature key pair of the profile.
-         * @property publicKeyEncryptionKeyPair The public key encryption key pair of the profile.
-         * @property securityState The security state of the profile (nullable).
-         */
-        @Serializable
-        data class AccountKeys(
-            @SerialName("signatureKeyPair")
-            val signatureKeyPair: SignatureKeyPair?,
-
-            @SerialName("publicKeyEncryptionKeyPair")
-            val publicKeyEncryptionKeyPair: PublicKeyEncryptionKeyPair,
-
-            @SerialName("securityState")
-            val securityState: SecurityState?,
-        ) {
-
-            /**
-             * Represents a signature key pair in the vault response.
-             *
-             * @property wrappedSigningKey The wrapped signing key of the signature key pair.
-             * @property verifyingKey The verifying key of the signature key pair.
-             */
-            @Serializable
-            data class SignatureKeyPair(
-                @SerialName("wrappedSigningKey")
-                val wrappedSigningKey: String,
-
-                @SerialName("verifyingKey")
-                val verifyingKey: String,
-            )
-
-            /**
-             * Represents a public key encryption key pair in the vault response.
-             *
-             * @property wrappedPrivateKey The wrapped private key of the public key encryption key
-             * pair.
-             * @property publicKey The public key of the public key encryption key pair.
-             * @property signedPublicKey The signed public key of the public key encryption key pair
-             * (nullable).
-             */
-            @Serializable
-            data class PublicKeyEncryptionKeyPair(
-                @SerialName("wrappedPrivateKey")
-                val wrappedPrivateKey: String,
-
-                @SerialName("publicKey")
-                val publicKey: String,
-
-                @SerialName("signedPublicKey")
-                val signedPublicKey: String?,
-            )
-
-            /**
-             * Represents security state in the vault response.
-             *
-             * @property securityState The security state of the profile.
-             * @property securityVersion The security version of the profile.
-             */
-            @Serializable
-            data class SecurityState(
-                @SerialName("securityState")
-                val securityState: String,
-
-                @SerialName("securityVersion")
-                val securityVersion: Int,
-            )
-        }
     }
 
     /**
